refactor(box): convert to typescript
