+++ src/components/box/box.config.ts
