--- src/components/box/index.js
@@ -1,1 +1,0 @@
-export { default } from "./box.component";
