+++ src/components/box/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./box.component";
+export type { BoxProps } from "./box.component";
